Add a test file specifically for websubmit features
* Add a shared fixture to start the server
